Delete TabLink, move convertForAria to stringUtils.ts, update imports accordingly
