+++ packages/manager/src/utilities/stringUtils.ts
@@ -28,3 +28,10 @@ export const truncateAndJoinList = (
 export const wrapInQuotes = (s: string) => '"' + s + '"';
 
 export const isNumeric = (s: string) => /^\d+$/.test(s);
+
+export const convertForAria = (str: string) => {
+  return str
+    .trim()
+    .toLowerCase()
+    .replace(/([^A-Z0-9]+)(.)/gi, (match, p1, p2) => p2.toUpperCase());
+};
